fix(platform): cross-build failures on ppc64el and older mingw
CI surfaced two cross-build issues that don't appear under amd64
glibc / current mingw:

  ppc64el (musl):
    platform/posix/process.c:161,169,339 -- ioctl request constants
    (TIOCSCTTY, TIOCSWINSZ) on ppc64 / mips have bit 31 set.  musl
    declares ioctl as POSIX-strict "int ioctl(int fd, int request,
    ...)", so the unsigned-long constant gets implicitly converted
    to int and the compiler errors out under -Werror=overflow.
    Cast each request to (int) at the call site -- the kernel
    interprets the bits the same way, the conversion is now
    explicit, and glibc (where ioctl takes unsigned long) is
    unaffected.

  windows-amd64 / -i386 / -arm64 (older mingw):
    platform/win/process.c:207 -- the SDK has the ConPTY function
    declarations and HPCON typedef but lacks the proc-thread
    attribute number used to bind a pseudo-console at CreateProcess
    time.  Define PROC_THREAD_ATTRIBUTE_PSEUDOCONSOLE inline as
    0x20016 -- the value of ProcThreadAttributeValue(22, FALSE,
    TRUE, FALSE) -- guarded by #ifndef so it's a no-op on newer
    SDKs that already provide it.

No behaviour change; just makes the CI matrix green.

+++ platform/posix/process.c
@@ -151,14 +151,20 @@ int process_start(struct process *proc)
 					  pty_slave_path);
 				_exit(EXIT_FAILURE);
 			}
-			ioctl(slave, TIOCSCTTY, 0);
+			/* Cast the request to int because musl declares
+			 * ioctl as `int ioctl(int fd, int request, ...)`
+			 * (POSIX-strict), and on ppc64 / mips the kernel
+			 * encodes some request constants with bit 31 set,
+			 * tripping -Werror=overflow at the implicit
+			 * unsigned long → int conversion. */
+			ioctl(slave, (int)TIOCSCTTY, 0);
 			if (proc->pty_rows > 0 && proc->pty_cols > 0) {
 				struct winsize ws;
 
 				memset(&ws, 0, sizeof ws);
 				ws.ws_row = (unsigned short)proc->pty_rows;
 				ws.ws_col = (unsigned short)proc->pty_cols;
-				ioctl(slave, TIOCSWINSZ, &ws);
+				ioctl(slave, (int)TIOCSWINSZ, &ws);
 			}
 			dup2(slave, STDIN_FILENO);
 			dup2(slave, STDOUT_FILENO);
@@ -336,7 +342,7 @@ int pty_resize(struct process *proc, int rows, int cols)
 	memset(&ws, 0, sizeof ws);
 	ws.ws_row = (unsigned short)rows;
 	ws.ws_col = (unsigned short)cols;
-	if (ioctl(proc->in, TIOCSWINSZ, &ws) < 0)
+	if (ioctl(proc->in, (int)TIOCSWINSZ, &ws) < 0)
 		return -1;
 	return 0;
 }
